Improved re/connectivity issues in the chat
I made it so that when two clients are in a chatroom and client A disconnects, it will attempt to re-establish connection to the server. Any messages sent while waiting to reconnect will be stored locally. Once connection is re-established, the messages will be automatically sent to the server so that client B can now see the messages client A sent while he was disconnected from the internet.